Still active? Sent a message to you on reddit /u/CoderWhoLikesDBFZ_2
Was wondering if there were any plans to update this tool for the latest patch, I tried it out and it doesn't seem to be working. If you are no longer active and have no plans on updating it, I sent you a PM about a proposition for your time. Thanks!